more robust emails
