+++ Core/Features/Moderation/UserReportEmailPayload.cs
@@ -1,0 +1,11 @@
+namespace LaHistoricalMarkers.Core.Features.Moderation
+{
+    public class UserReportEmailPayload
+    {
+        public int MarkerId { get; set; }
+
+        public string Report { get; set; }
+
+        public string Otp { get; set; }
+    }
+}
